feat: simplification de la gestion des erreurs et du chargement
Réduction du code redondant dans `EtudiantsAjoutView` en supprimant des fonctionnalités inutilisées comme la gestion locale des semestres et des années. Ajout d'une gestion des erreurs avec `ErrorView` dans les composants concernés pour améliorer l'UX.
